chore: game api endpoint example
