+++ gametrail/api/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework.serializers import ModelSerializer
+from gametrail.models import Game
+
+class GameSerializer(ModelSerializer):
+    class Meta:
+        model = Game
+        fields = '__all__'
